Updated file structure - code is NOT functional yet
