+++ lib/include/trajectory_xml_parser.h
