+++ lib/include/ttt_moves_trajectories.h
