+++ lib/src/board_sensor/board_motion_detection.cpp
@@ -1,6 +1,6 @@
 
-#include "ttt/tictactoe_utils.h"
-#include "ttt/board_motion_detection.h"
+#include "../../include/tictactoe_utils.h"
+#include "../../include/board_motion_detection.h"
 
 namespace ttt
 {
