+++ lib/src/tictactoe/msg/_PlaceTokenResult.py
