+++ lib/src/trajectory_player/trajectory_player.cpp
