+++ lib/src/trajectory_xml_parser/concat_trajectory_xml_files.cpp
