+++ lib/src/trajectory_xml_parser/trajectory_from_raw_to_xml.cpp
