+++ lib/src/trajectory_xml_parser/trajectory_xml_parser.cpp
