+++ package.xml
@@ -46,6 +46,7 @@
   <author>Álvaro Castro González [acgonzal@ing.uc3m.es]</author>
   <maintainer email="olivier.mangin@yale.edu">Olivier Mangin</maintainer>
   <maintainer email="alessandro.roncone@yale.edu">Alessandro Roncone</maintainer>
+  <maintainer email="johnamadeo.daniswara@yale.edu">John Amadeo Daniswara</maintainer>
   <license>GNU GPL v2</license>
 
   <buildtool_depend>catkin</buildtool_depend>
@@ -56,11 +57,8 @@
   <depend>rosconsole</depend>
   <depend>actionlib</depend>
   <depend>actionlib_msgs</depend>
-  <depend>ttt_board_sensor</depend>
   <depend>baxter_core_msgs</depend>
   <depend>sound_play</depend>
-
-  ## ttt_board_sensor
   <depend>sensor_msgs</depend>  
   <depend>opencv2</depend>  
   <depend>cv_bridge</depend>
